// src/lib/api/types.gen.ts
export type PaymentCategoryType =
  | "RENT"
  | "UTILITIES"
  | "LOAN"
  | "FOOD"
  | "TRANSPORT"
  | "INSURANCE"
  | "SUBSCRIPTION"
  | "SALARY"
  | "FREELANCE"
  | "OTHER";